Passing Props
created a function Component Square() which returned the previous button. Called said component using JSX Syntax in the Board function to show the reusing of a component. Modifed the function Square() to accept a props that we created call value. In order for props to be called properly we needed to wrap the property in curly braces to escape the JSX back into Javascript.  Assigned a value attribute to each of the component called as the final step.
git status