Bringing State monitoring up to the parent
moved the state tracking and the handleClick to the parent component Board() Modifed the handleClick function to accept i in order to change the respective squares clicked. Maintain immuntability by creating a copy and modifying the copy only to push back as the new state. Initialized squares with an array of 9 nulls. Modified the Square component to accept props of value, onSquareClick. added attribute of value and onSquareClick to the jSX calling of the square component. Note the onSquareClick attribute is an arrow function to not call the handleClick() on load as it would result in an endless loop of re-rendering